+++ src/main/java/com/WebSite/demo/web/LessonController.java
@@ -11,12 +11,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 public class LessonController {
     /**
      * show lesson to user
-     * @param id
-     * @param model
-     * @return
+     * @param id lesson id
+     * @return lesson
      */
     @GetMapping("/lesson/{id}")
-    public String showLesson(@PathVariable("id") int id, Model model){
+    public String showLesson(@PathVariable("id") long id, Model model){
         Lesson lesson = LessonDao.findLessonById(id);
         model.addAttribute(lesson);
         return "lesson";
